+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Link`: fixed behaviour of `style` when `theme.color` is nullish
 - add `MultiBanner` component
 - add `NoScript` component
+- add `ResponsiveFlex` component
 
 ## `@svizzle/atlas` v0.6.0 (next)